pannuke_inference_dist/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
Pillow>=8.0.0
tqdm>=4.60.0
timm>=0.4.12
albumentations>=1.0.0
opencv-python>=4.5.0
